// frontend/app/src/app/models/korisnik.ts
export class Korisnik{
    kor_ime: string;
    lozinka: string;
    ime: string;
    prezime: string;
    adresa: string;
    telefon: string;
    mejl: string;
    tip: string;
    slika: string;
    status: string;
    licenca: string;
    specijalizacija: string;
    ordinacija: string;
    godisnji_pocetak: string;
    godisnji_kraj: string;
    slobodan_dan: string[]
}